Remove handshake
