+++ oxidation/libparsec/crates/protocol/src/error.rs
@@ -2,35 +2,6 @@
 
 use thiserror::Error;
 
-use libparsec_types::prelude::*;
-
-use crate::{ApiVersion, HANDSHAKE_CHALLENGE_SIZE};
-
-#[derive(Error, Debug, Clone, PartialEq)]
-pub enum HandshakeError {
-    #[error("Handshake invalid answer (bad signature or challenge value)")]
-    FailedChallenge,
-    #[error("Handshake bad administration token")]
-    BadAdministrationToken,
-    #[error("Handshake bad identity")]
-    BadIdentity,
-    #[error("Handshake organization expired")]
-    OrganizationExpired,
-    #[error("Handshake rvk mismatch")]
-    RVKMismatch,
-    #[error("Handshake revoked device")]
-    RevokedDevice,
-    #[error("Handshake out of ballpark {0:x?}")]
-    OutOfBallpark(ChallengeDataReport),
-    #[error("No overlap between client API versions {client_versions:?} and backend API versions {backend_versions:?}")]
-    APIVersion {
-        client_versions: Vec<ApiVersion>,
-        backend_versions: Vec<ApiVersion>,
-    },
-    #[error("Handshake invalid message: {0}")]
-    InvalidMessage(&'static str),
-}
-
 #[derive(Error, Debug, Clone, PartialEq, Eq)]
 pub enum ProtocolError {
     #[error("Encoding error: {exc}")]
@@ -47,14 +18,4 @@ pub enum ProtocolError {
     NotHandled,
     #[error("{exc}")]
     BadRequest { exc: String },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ChallengeDataReport {
-    pub challenge: [u8; HANDSHAKE_CHALLENGE_SIZE],
-    pub supported_api_versions: Vec<ApiVersion>,
-    pub backend_timestamp: DateTime,
-    pub client_timestamp: DateTime,
-    pub ballpark_client_early_offset: f64,
-    pub ballpark_client_late_offset: f64,
 }
